+++ apps/roam/src/components/ExportGithub.tsx
@@ -12,7 +12,7 @@ import apiGet from "roamjs-components/util/apiGet";
 import apiPost from "roamjs-components/util/apiPost";
 import { getNodeEnv } from "roamjs-components/util/env";
 import getExtensionApi from "roamjs-components/util/extensionApiContext";
-import { setSetting } from "~/utils/settings";
+import { setSetting } from "~/utils/extensionSettings";
 
 type UserReposResponse = {
   data: [
